Move Exhibit card in its own folder
