--- src/components/Feed/index.js
@@ -1,2 +1,0 @@
-import Feed from './Feed';
-export default Feed;
